Rename CosmeticService to BeautyCare

+++ src/main/java/com/marcellkrausz/appointmentreserve/services/implementations/AddressServiceImpl.java
@@ -9,6 +9,7 @@ import com.marcellkrausz.appointmentreserve.models.Address;
 import com.marcellkrausz.appointmentreserve.repositories.AddressRepository;
 import com.marcellkrausz.appointmentreserve.services.AddressService;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.HashSet;
@@ -33,6 +34,10 @@ public class AddressServiceImpl implements AddressService {
     public Set<AddressDto> getAllAddress() {
         Set<Address> addresses = new HashSet<>();
         addressRepository.findAll().iterator().forEachRemaining(addresses::add);
+        if (addresses.isEmpty()) {
+            throw new AddressNotFoundException("Addresses not found in database.");
+        }
+
         return addressToAddressDto.convertSet(addresses);
     }
 
@@ -40,8 +45,9 @@ public class AddressServiceImpl implements AddressService {
     public AddressDto getAddressById(Long id) {
         Optional<Address> addressOptional = addressRepository.findById(id);
         if (addressOptional.isEmpty()) {
-            throw new AddressNotFoundException("Address not found");
+            throw new AddressNotFoundException("Address not found.");
         }
+
         return addressToAddressDto.convert(addressOptional.get());
     }
 
@@ -55,6 +61,11 @@ public class AddressServiceImpl implements AddressService {
 
     @Override
     public void deleteAddressById(Long id) {
+        Optional<Address> addressOptional = addressRepository.findById(id);
+        if (addressOptional.isEmpty()) {
+            throw new AddressNotFoundException("Address not found.");
+        }
+
         addressRepository.deleteById(id);
         log.debug("Deleted address id: " + id);
     }
